requirements.txt
streamlit
pandas
psycopg2-binary
streamlit-authenticator
bcrypt
PyYAML
python-dotenv